Move proposal fixtures to test support
